YOLO Trader/requirements.txt
alpaca-trade-api
pandas
yfinance
numpy
scikit-learn
backtrader
dash
plotly
python-dotenv